+++ src/config.rs
@@ -1,6 +1,7 @@
 use crate::{json, JsonBookmarks, Settings};
 use anyhow::{anyhow, Context};
 use log::{debug, trace};
+use rlimit::Resource;
 use std::{
     env,
     fs::{self, File},
@@ -69,6 +70,17 @@ impl Config {
         }
 
         let settings = Settings::init(&settings_path)?;
+
+        let (soft_limit, hard_limit) = rlimit::getrlimit(Resource::NOFILE)?;
+        debug!("Soft and hard limit for file descriptors: {soft_limit} and {hard_limit}");
+
+        // `max_concurrent_requests` is bounded by how fast we can write the
+        // content, i.e. by the maximum number of open file descriptors.
+        let max_file_descriptors = settings.max_concurrent_requests as u64;
+
+        if soft_limit < max_file_descriptors || hard_limit < max_file_descriptors {
+            rlimit::setrlimit(Resource::NOFILE, max_file_descriptors, max_file_descriptors)?;
+        }
 
         if !target_bookmark_path.exists() {
             debug!(